사이트 제작 Item Page 2.1
// TODO
>> 작성한 상품의 상세페이지 작성 시작!
>> ItemCard 와 Pagination 모듈화하여 분리.

// 네이버, 카카오 간편로그인 구현해야함!

+++ src/main/java/com/lec/spring/service/ItemService.java
@@ -25,7 +25,7 @@ public interface ItemService {
     // Find Item By UserKey
     List<Item> findByUserKey(Long userKey);
 
-    int write(Item item, List<ItemImage> imageList);
+    Long write(Item item, List<ItemImage> imageList);
 
     // Item's Detail
     Item detail(Long itemKey);
